Moving interfaces from `src/schema` to `src/`
