+++ package.json
@@ -25,7 +25,7 @@
     "prestart": "node scripts/index-examples",
     "start": "npm run watch & browser-sync start --server --files 'vega-lite.js' --index 'test-gallery.html'",
     "poststart": "rm examples/all-examples.json",
-    "schema": "typescript-json-schema --required true src/schema/schema.ts Spec > vega-lite-schema.json",
+    "schema": "typescript-json-schema --required true src/spec.ts Spec > vega-lite-schema.json",
     "images": "scripts/generate-images.sh",
     "presite": "tsc && npm run build && bower install",
     "site": "bundle exec jekyll serve --incremental",
