+++ src/compile/data.ts
@@ -1,8 +1,8 @@
 import * as vlFieldDef from '../fielddef';
 import {extend, keys, vals, reduce, contains} from '../util';
 import {Model} from './Model';
-import {FieldDef} from '../schema/fielddef.schema';
-import {VgData} from '../schema/vega.schema';
+import {FieldDef} from '../fielddef';
+import {VgData} from '../vega.schema';
 import {StackProperties} from './stack';
 import {ScaleType} from '../enums';
