+++ src/data.ts
@@ -1,8 +1,19 @@
 /*
  * Constants and utilities for data.
  */
+import {Type} from './type';
+import {DataFormat} from './enums';
 
-import {NOMINAL, QUANTITATIVE, TEMPORAL} from './type';
+export interface Data {
+
+  formatType?: DataFormat;
+  url?: string;
+  /**
+   * Pass array of objects instead of a url to a file.
+   */
+  values?: any[];
+}
+
 
 export const SUMMARY = 'summary';
 export const SOURCE = 'source';
@@ -10,11 +21,11 @@ export const STACKED_SCALE = 'stacked_scale';
 export const LAYOUT = 'layout';
 
 /** Mapping from datalib's inferred type to Vega-lite's type */
-// TODO: ALL_CAPS
+// TODO: consider if we can remove
 export const types = {
-  'boolean': NOMINAL,
-  'number': QUANTITATIVE,
-  'integer': QUANTITATIVE,
-  'date': TEMPORAL,
-  'string': NOMINAL
+  'boolean': Type.NOMINAL,
+  'number': Type.QUANTITATIVE,
+  'integer': Type.QUANTITATIVE,
+  'date': Type.TEMPORAL,
+  'string': Type.NOMINAL
 };
