--- src/schema/config.cell.schema.ts
@@ -1,29 +1,0 @@
-export interface CellConfig {
-  width?: number;
-  height?: number;
-
-  clip?: boolean;
-
-  // FILL_STROKE_CONFIG
-  /**
-   * @format color
-   */
-  fill?: string;
-  fillOpacity?: number;
-  stroke?: string;
-  strokeWidth?: number;
-  strokeOpacity?: number;
-  strokeDash?: number;
-  /** The offset (in pixels) into which to begin drawing with the stroke dash array. */
-  strokeDashOffset?: number;
-}
-
-export const defaultCellConfig: CellConfig = {
-  width: 200,
-  height: 200
-};
-
-export const defaultFacetCellConfig: CellConfig = {
-  stroke: '#ccc',
-  strokeWidth: 1
-};
