+++ src/spec.ts
@@ -1,13 +1,37 @@
 /* Utilities for a Vega-Lite specificiation */
 
-import {FieldDef} from './schema/fielddef.schema';
-import {Spec} from './schema/schema';
+import {FieldDef} from './fielddef';
+// Package of defining Vega-lite Specification's json schema
 
-import {Model} from './compile/Model';
+import {Config} from './config';
+import {Data} from './data';
+import {Encoding} from './encoding';
+import {Mark} from './mark';
+import {Transform} from './transform';
+
 import {COLOR, SHAPE} from './channel';
 import * as vlEncoding from './encoding';
 import {BAR, AREA} from './mark';
 import {duplicate} from './util';
+
+/**
+ * Schema for Vega-Lite specification
+ * @required ["mark", "encoding"]
+ */
+export interface Spec {
+  /**
+   * A name for the specification. The name is used to annotate marks, scale names, and more.
+   */
+  name?: string;
+  description?: string;
+  data?: Data;
+  transform?: Transform;
+  mark: Mark;
+  encoding: Encoding;
+  config?: Config;
+}
+
+
 
 // TODO: add vl.spec.validate & move stuff from vl.validate to here
 
@@ -23,7 +47,7 @@ export function fieldDefs(spec: Spec): FieldDef[] {
 
 export function getCleanSpec(spec: Spec): Spec {
   // TODO: move toSpec to here!
-  return new Model(spec).toSpec(true);
+  return spec;
 }
 
 export function isStack(spec: Spec): boolean {
