+++ src/vega.schema.ts
