+++ src/vl.ts
@@ -4,7 +4,6 @@ import * as vlData from './data';
 import * as vlEncoding from './encoding';
 import * as vlFieldDef from './fielddef';
 import * as vlCompile from './compile/compile';
-import * as vlSchema from './schema/schema';
 import * as vlShorthand from './shorthand';
 import * as vlSpec from './spec';
 import * as vlTimeUnit from './timeunit';
@@ -18,7 +17,6 @@ export const compile = vlCompile.compile;
 export const data = vlData;
 export const encoding = vlEncoding;
 export const fieldDef = vlFieldDef;
-export const schema = vlSchema;
 export const shorthand = vlShorthand;
 export const spec = vlSpec;
 export const timeUnit = vlTimeUnit;
