+++ tsconfig.json
@@ -32,6 +32,7 @@
     "site/static/main.ts",
     "site/static/test-gallery.ts",
     "src/aggregate.ts",
+    "src/axis.ts",
     "src/bin.ts",
     "src/channel.ts",
     "src/compile/axis.ts",
@@ -53,32 +54,23 @@
     "src/compile/scale.ts",
     "src/compile/stack.ts",
     "src/compile/time.ts",
+    "src/config.ts",
     "src/data.ts",
     "src/encoding.ts",
     "src/enums.ts",
     "src/fielddef.ts",
+    "src/legend.ts",
     "src/mark.ts",
-    "src/schema/axis.schema.ts",
-    "src/schema/bin.schema.ts",
-    "src/schema/config.cell.schema.ts",
-    "src/schema/config.facet.schema.ts",
-    "src/schema/config.mark.schema.ts",
-    "src/schema/config.schema.ts",
-    "src/schema/data.schema.ts",
-    "src/schema/encoding.schema.ts",
-    "src/schema/fielddef.schema.ts",
-    "src/schema/legend.schema.ts",
-    "src/schema/scale.schema.ts",
-    "src/schema/schema.ts",
-    "src/schema/sort.schema.ts",
-    "src/schema/transform.schema.ts",
-    "src/schema/vega.schema.ts",
+    "src/scale.ts",
     "src/shorthand.ts",
+    "src/sort.ts",
     "src/spec.ts",
     "src/timeunit.ts",
+    "src/transform.ts",
     "src/type.ts",
     "src/util.ts",
     "src/validate.ts",
+    "src/vega.schema.ts",
     "src/vl.ts",
     "test/compile/axis.test.ts",
     "test/compile/common.test.ts",
